+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Improve performance of two-layer operations using `nb_parallel=1` (#692)
 - Add checks on invalid values in `ConfigOptions` (#711)
 - Add worker_type used to progress logging (#715)
+- Write gdal log files to `GFO_TMPDIR` if specified (#727)
 
 ## 0.10.2 (2025-08-20)