+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Improvements
 
+- Improve performance of most operations by using a direct gpkg to gpkg append via
+  sqlite where possible (#728)
 - Ensure that the featurecount is properly cached in GPKG files, also for older GDAL
   versions + small refactor (#693)
 - Improve performance of two-layer operations using `nb_parallel=1` (#692)