+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Ensure that the featurecount is properly cached in GPKG files, also for older GDAL
   versions + small refactor (#693)
 - Improve performance of two-layer operations using `nb_parallel=1` (#692)
+- Reduce commited memory being allocated (#739)
 - Add checks on invalid values in `ConfigOptions` (#711)
 - Add worker_type used to progress logging (#715)
 - Write gdal log files to `GFO_TMPDIR` if specified (#727)